+++ src/main/java/org/bytedeco/javacv/OpenCVFrameConverter.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2015 Samuel Audet
+ * Copyright (C) 2015-2018 Samuel Audet
  *
  * Licensed either under the Apache License, Version 2.0, or (at your option)
  * under the terms of the GNU General Public License as published by
@@ -23,6 +23,7 @@
 package org.bytedeco.javacv;
 
 import java.nio.Buffer;
+import java.nio.ByteBuffer;
 import org.bytedeco.javacpp.BytePointer;
 import org.bytedeco.javacpp.Pointer;
 
@@ -38,6 +39,7 @@ import static org.bytedeco.javacpp.opencv_core.*;
 public abstract class OpenCVFrameConverter<F> extends FrameConverter<F> {
     IplImage img;
     Mat mat;
+    org.opencv.core.Mat orgOpenCvCoreMat;
 
     public static class ToIplImage extends OpenCVFrameConverter<IplImage> {
         @Override public Frame convert(IplImage img) { return super.convert(img); }
@@ -47,6 +49,11 @@ public abstract class OpenCVFrameConverter<F> extends FrameConverter<F> {
     public static class ToMat extends OpenCVFrameConverter<Mat> {
         @Override public Frame convert(Mat mat) { return super.convert(mat); }
         @Override public Mat convert(Frame frame) { return convertToMat(frame); }
+    }
+
+    public static class ToOrgOpenCvCoreMat extends OpenCVFrameConverter<org.opencv.core.Mat> {
+        @Override public Frame convert(org.opencv.core.Mat mat) { return super.convert(mat); }
+        @Override public org.opencv.core.Mat convert(Frame frame) { return convertToOrgOpenCvCoreMat(frame); }
     }
 
     public static int getFrameDepth(int depth) {
@@ -78,7 +85,7 @@ public abstract class OpenCVFrameConverter<F> extends FrameConverter<F> {
         return img != null && frame != null && frame.image != null && frame.image.length > 0
                 && frame.imageWidth == img.width() && frame.imageHeight == img.height()
                 && frame.imageChannels == img.nChannels() && getIplImageDepth(frame.imageDepth) == img.depth()
-                && new Pointer(frame.image[0]).address() == img.imageData().address()
+                && new Pointer(frame.image[0].position(0)).address() == img.imageData().address()
                 && frame.imageStride * Math.abs(frame.imageDepth) / 8 == img.widthStep();
     }
     public IplImage convertToIplImage(Frame frame) {
@@ -127,7 +134,7 @@ public abstract class OpenCVFrameConverter<F> extends FrameConverter<F> {
         return mat != null && frame != null && frame.image != null && frame.image.length > 0
                 && frame.imageWidth == mat.cols() && frame.imageHeight == mat.rows()
                 && frame.imageChannels == mat.channels() && getMatDepth(frame.imageDepth) == mat.depth()
-                && new Pointer(frame.image[0]).address() == mat.data().address()
+                && new Pointer(frame.image[0].position(0)).address() == mat.data().address()
                 && frame.imageStride * Math.abs(frame.imageDepth) / 8 == (int)mat.step();
     }
     public Mat convertToMat(Frame frame) {
@@ -153,6 +160,63 @@ public abstract class OpenCVFrameConverter<F> extends FrameConverter<F> {
             frame.imageChannels = mat.channels();
             frame.imageStride = (int)mat.step() * 8 / Math.abs(frame.imageDepth);
             frame.image = new Buffer[] { mat.createBuffer() };
+        }
+        frame.opaque = mat;
+        return frame;
+    }
+
+    static boolean isEqual(Frame frame, org.opencv.core.Mat mat) {
+        return mat != null && frame != null && frame.image != null && frame.image.length > 0
+                && frame.imageWidth == mat.cols() && frame.imageHeight == mat.rows()
+                && frame.imageChannels == mat.channels() && getMatDepth(frame.imageDepth) == mat.depth()
+                && new Pointer(frame.image[0].position(0)).address() == mat.dataAddr();
+    }
+    public org.opencv.core.Mat convertToOrgOpenCvCoreMat(Frame frame) {
+        if (frame == null || frame.image == null) {
+            return null;
+        } else if (frame.opaque instanceof org.opencv.core.Mat) {
+            return (org.opencv.core.Mat)frame.opaque;
+        } else if (!isEqual(frame, mat)) {
+            int depth = getMatDepth(frame.imageDepth);
+            orgOpenCvCoreMat = depth < 0 ? null : new org.opencv.core.Mat(frame.imageHeight, frame.imageWidth,
+                    CV_MAKETYPE(depth, frame.imageChannels), new BytePointer(new Pointer(frame.image[0].position(0)))
+                            .capacity(frame.image[0].capacity() * Math.abs(frame.imageDepth) / 8).asByteBuffer());
+        }
+        return orgOpenCvCoreMat;
+    }
+    public Frame convert(final org.opencv.core.Mat mat) {
+        if (mat == null) {
+            return null;
+        } else if (!isEqual(frame, mat)) {
+            frame = new Frame();
+            frame.imageWidth = mat.cols();
+            frame.imageHeight = mat.rows();
+            frame.imageDepth = getFrameDepth(mat.depth());
+            frame.imageChannels = mat.channels();
+            frame.imageStride = (int)mat.step1();
+            ByteBuffer byteBuffer = new BytePointer() { { address = mat.dataAddr(); } }.capacity(mat.rows() * mat.step1() * mat.elemSize1()).asByteBuffer();
+            switch (mat.depth()) {
+                case CV_8U:
+                case CV_8S:
+                    frame.image = new Buffer[] { byteBuffer };
+                    break;
+                case CV_16U:
+                case CV_16S:
+                    frame.image = new Buffer[] { byteBuffer.asShortBuffer() };
+                    break;
+                case CV_32F:
+                    frame.image = new Buffer[] { byteBuffer.asFloatBuffer() };
+                    break;
+                case CV_32S:
+                    frame.image = new Buffer[] { byteBuffer.asIntBuffer() };
+                    break;
+                case CV_64F:
+                    frame.image = new Buffer[] { byteBuffer.asDoubleBuffer() };
+                    break;
+                default:
+                    frame.image = null;
+                    break;
+            }
         }
         frame.opaque = mat;
         return frame;